Slimming down cron-tool
